move api to auth/api
